refactor(@formatjs/editor): use visual component module naming
