+++ packages/editor/vrt/editor.browser.spec.ts
@@ -1,5 +1,5 @@
 import {expect, test} from '@playwright/test'
-import type {Editable} from './editor.fixture.js'
+import type {Editable} from './editor.visual.js'
 
 test('mounted editor retains translation while its provider direction changes', async ({
   mount,
